[FIX] issue when importing products from CSV with procurement_type attribute
There is an issue in the framework (I think) which will break the previous
version of the code in CSV imports: the get_product_procurement_type method (in
product_template) is called with an instance of product_product for self, and
the call to super() crashed. This redesign works around the issue by extracting
get_product_procurement_type to a function and thus avoiding self altogether.