fix use of @api.one
`@api.one` will concatenate the return value of the function in a list for all
records in the browse_record, which requires getting elements from that list
afterwards.

In case the method is meant to be call on an individual record, it is better to
use `@api.multi` and a call to ensure_one to make sure there is a
single id in `ids`.