check for consistency between procurement and PO
This check is done when the procurement is processed, which happens when
the Sale Order is validated. We cannot easily check before validation
because we need to know what destination location has been chosen in the
procurement. Just checking the route selected in the Sale Order Line is
too restrictive because the Route can be chosen in many other places.