extend sale_exception_nostock to by_line case
In v7, an extra module allowed to select the source location in every
sale order line. In v8 standard, you can select an optional Route in
every line and get the same result.

For that reason, instead of porting the old
sale_exception_nostock_by_line to v8, I extended the
sale_exception_nostock module so that it takes care of the Route
selected if the order line if there is one.