Add automatic relation of SO line and logisitc route based on sourced_by field.
Wee look for the PO related to current SO line by the sourced_by fields.

If the PO has a destination location with usage "customer" we apply the dropshipping route to SO line.

If the PO has a destination location with usage "internal" we apply the make to order route to SO line.