adapt the test to test the new onchange
We are now green again, and the test looks more readable to me.

On the other hand, we lost the test for the warning in case of
inconsistent PO destination as SOL route.

I am not convinced this check is still necessary. It does no harm still:
we seem to always follow the happy path.