Remove the filter on incoming pickings from the purchase orders
The pickings generated by purchase orders are changed to outgoing pickings when
the order's flow is drop shipping.  The 'view_picking' function gets the action
to show the pickings from the XML action. This action has a domain "('type',
'=', 'in')", that prevents to see the drop shipping pickings as they have the
type 'out'.

This commit totally removes the "('type', '=', 'in')" condition from the
action: it is useless because we already filter on the pickings attached to the
purchase order (on the 'picking_ids' field).