Unit test to assert right state of order lines
flake8

PEP

There was another problem, the old order lines became the new order lines and vice versa, but the old procurement order referred to the old order line and thus now to the new order line, so that when creating a shipment, the old procurement order was used containing the wrong lines. Long story short, instead of also switching the procurement order around, a better strategy is to just swap the lines of the old and new order completely.

Style improvements