Rename line_ship to sale_allotment and improve it for OCA adoption
Keep the same parameters for all the pickings
Directly use the sale line's allotment. The procurement's partner is used for drop shipping, we should probably not hijack it.
Add the allotment partner to the sale line's treeview. Those who don't use other advanced features can't see the field otherwise because it's only in the form view.
Create allotment partners as children of the customer, with type='other' by default. Creating them as type 'delivery' by default would be error-prone, since they would always be proposed for the next orders.
Conform to OCA conventions
Add tests for sale_allotment

NB: We don't group allotments on the same picking if the partner uses the parent adress. It was suggested but actually a bit naive given the very loose coupling we now have between stock moves and pack operations. It turned out to be mostly unworkable, because stock workers don't have the dispatch information at the right place anymore.