[FIX] sale_order_type: Keep sale type on generated invoices
Steps to reproduce:

* Create a partner with sale type A.
* Create an SO for that partner with sale type B.
* Generate the invoice for that SO.
* Generated invoice will have type A instead of type B.

This is caused by the fact that after creating invoices, there's a process in core
called `_finalize_invoices` that calls the onchange of the partner for bringing
other missing data, but as there's no filter, it overwrites all fields that are
populated through `_onchange_partner_id` successive overwrites.

We fix it saving previous values before calling the method, and comparing afterwards
if it has been changed, restoring it if so. It's an ugly code, but there's no
better way (mangling with values passed by context for disabling the onchange will
populate other environments that will have undesired effects).