[IMP] sale_order_type: run onchanges on create
When an order was created outside UI (e.g.: website sale) the order type
values weren't properly fetched. We run them on create when no type is
given asumming that in that case the invoker didn't take this behavior
into account.

Also, we get rid of the default value for the field and keep although we
default to it in case none is given by the partner.