[IMP][FIX] Update picking note on all SO related Pickings
When having multi step rules, such as pick-pack, the picking notes
are now copied to all related Pickings, and not only to the Delivery
Order.

An option was added to allow changing the picking note once the SO is
confirmed, and have the note updates also on the Pickings.
This is useful in case we have additional information for the warehouse
after the customer confirms the order.