[FIX] 11.0 Sale Fixed Discount
Fix
===
`_onchange_discount` method was shadowing the original _onchange_discount
and all its field triggers had been replaced by `discount`. As the new method's
trigger is the field `discount`, the name should be `_onchange_discount`.
But the name is already taken. Hence `_onchange_discount2`. I'm not aware of
a better naming convention in cases like this.

Also, the way the method was implemented, it was impossible to change the
discount manually if the pricelist had a discount, because that would
trigger the `_onchange_discount`, call to super, and get the discount
from the pricelist.

Improve
=======
Override orginal `_onchange_discount` in order to get the value for the
fixed discount in case the price rule method is `fixed`. It basically
repeats what the original method does, so not very DRYish. I don't know
if there is a better approach (I can only think of monkey path, but I
don't like it either)