Fix test checking onchanges
The test was calling the onchanges manually to test their result, but as
some onchanges must be called in cascade to have a correct result, the
test tells more about itself, if it calls the onchanges in the correct
order or not.
Using Form will test them correctly in cascade automatically.

Signed-off-by: Carmen Bianca BAKKER <carmen@coopiteasy.be>