[13.0][FIX] sale_stock_delivery_addres: order_partner_id not being stored.
Include the related stored field `order_partner_id` in the views
has the side effect of preventing it to be stored, even when it
is invisible. ORM issues appart it is better to just remove it
in this case, as the field is not needed anymore in the view by
this module.