[FIX] sale_coupon_advanced: counter forced reward
In order for it to work in all cases, instead of removing incorrectly
forced lines, now we check if line can be added and add that line before
it is expected to be used by sale_coupon module.

Because trigger points are different: for applying coupon and
recomputing promotions, had to inject this logic into multiple places.

Bugs fixed:
    - no longer adds unrelated programs reward products on SO when
        applying coupon code from forced program.
    - no longer adds products from forced promotions that need code to
        be entered, when recomputing promotions.
    - handling order line sequence generation, when it is called, when
        no lines exists.