[FIX] sale_coupon_multi_use: sale confirm (#20)
Coupons are marked as consumed twice. First time when coupon is added
and second when sale order is confirmed.

For that, multi-use coupon must not be marked as consumed when
confirming order (if it is still valid by multi-use rules).

Also had to redesign the way multi use coupons are checked on write,
because during confirmation it can update multiple coupons and old
implementation was able to handle one coupon at a time.

On top of that, simplified multi-use coupon handling. Now correct amount
is intercepted when discount_fixed_amount is computed for reward line.

This gives multiple advantages:
    - no need to adjust sale order line discount afterwards.
    - _compute_program_amount handles currency conversion, so multi-use
        consumed amount now also takes care of that too.
    - simpler consumption line creation and checking if coupon is fully
        consumed (no need to compare delta with discount that would need
        to be computed from sale order).

[FIX] sale_coupon_multi_use: action_cancel

Make sure related consumption lines are removed if coupon is being
reset (made valid again), e.g. SO is being cancelled.

[REF] sale_coupon_multi_use: LIC/naming/misc/lang