Apply cut-off time on commitment date
When a commitment date is set on the sale order, respect the date but
change the cut-off time. This will help to keep all scheduled datetimes
in line (e.g. if we want to merge them).

I added a lead time of 1 day in the test setup which is a usual configuration,
so I had to adapt all the other test results.