[FIX] Existing SO can't be cancelled after install.
If this module is installed in a environment where there are sales
orders not in draft state. The added field type_id is left null
then the order can't be cancelled.

A hook is provided to fill the type_id with data. It is a hook
but not a migration becouse it must run also at first install.