[13.0] [FIX] sale_order_lot_selection: Adjust problems with reservetion for the qty using a lot.
In some case you don't have enought qty in one lot for cover the initial demand the system reserved for the all the quants the qtty to complete the initial demand, but the asignation in the smls are
completed with the lot chossing in the sale order line. This inconsit produce a lot a problems with the units reserved and you cannnot cancel the all reservation.
To solve this situation i propose this PR to ensure that Lot in the sol is use for the qty that may have and  if you could use others lots with quantity to complete if it's posible by the qty avaliable the inicial demand.