[OU-FIX] sale_automatic_workflow: complete noupdate changes + model switch
The model of the filter automatic_workflow_validate_invoice_filter has
also changed and it's not included in the noupdate changes.

We also make a preventing model change in all defined filters for
avoiding that error, although the fields used in it may not exist in
the new model or the name has changed.

TT33685