[FIX] sale_order_line_price_history: Field order_partner_id is not updated when change partner_id
When manipulating the dependencies of related fields, we may need to replicate the relationship itself when the field is on the screen and not readonly.
Because if we force the depends for when they act as Onchange/NewId and the field is on the screen (as of v13 it is saved even if it is invisible without the need for force_save)
So in this case only the related was computed in "onchange mode" when the product was changed and since the field was defined in the line, the old value was saved, so the related field was not obtained from relation when save document.

TT41519

[FIX] sale_order_line_price_history: Field order_partner_id is not updated when change partner_id (Only JS)
TT41519