[FIX] sale_order_line_chained_move: Fix infinite recursion
The `__find_origin_moves` function doesn't take into account loops in
the chains of origin moves.  Fix this by adding a "visited" record set
which will make sure that any move that has already be expanded won't be
processed again.

This will allow updating the modules (`odoo-bin -u all`) even when there
is existing data which has cycles in their origin graph.  The error for
that case looked like this:

    INFO odoo odoo.modules.loading: Loading module sale_order_line_chained_move (185/225)
    ...
      File ".../thirdparty/odoo/addons/sale_order_line_chained_move/hooks.py", line 10, in __find_origin_moves
        all_moves |= __find_origin_moves(move.move_orig_ids)
      [Previous line repeated 978 more times]
      File ".../thirdparty/odoo/addons/sale_order_line_chained_move/hooks.py", line 9, in __find_origin_moves
        if move.move_orig_ids:
      ...
    RecursionError: maximum recursion depth exceeded