[MIG] sale_order_secondary_unit: migrate to v16.0
Some inherited fields are set to `precompute=True` for performance.

The test `test_secondary_uom_unit_price` was asserting a buggy behavior
from v15 that is fixed automatically in v16 by the migration. Thus, we
change the assertions. Now, when qtys change, the unit price is
recomputed. That's the upstream behavior. It makes more sense, because
you might have pricelists that depend on the qty. If you want to keep
that old behavior, install `sale_order_qty_change_no_recompute`.

@moduon MT-3490

Co-authored-by: Jairo Llopis <973709+yajo@users.noreply.github.com>