[IMP] sale_order_general_discount: allow to include specific products
This commit introduces a new setting that allows to set a domain
on the products that are meant to be included in the general discount.
Sale order lines with products that are excluded do not follow the
order's global discount rate.

If the filter is empty, all products are included, which preserves the
current behaviour of the module.