[ADD] sale_order_product_recommendation_packaging_default: glue module
This module extends the functionality of sale order product recommendations to
support gathering and specifying product packaging and to allow you to input
that info directly from the recommendations wizard.

To know how to use this module, follow the usage instructions of `sale_order_product_recommendation`.

If your database uses product packagings, you should be able to see and use 2 new columns in the recommendations wizard:

- *Packaging*
- *Packaging Qty.*

@moduon MT-3489