[FIX] sale_order_product_recommendation: freezegun hack
As seen in https://github.com/spulec/freezegun/issues/485, when decorating a class with `@freezegun.freeze_time()`, class cleanups are not properly executed.

For Odoo, this means that the cursor is not properly closed. Thus, if you happen to run a test for another module after this one that needs a Postgres-level lock that conflicts with this module's, it will fail.

This change does the same, but without decorating the class. Then, the cleanups work as expected.

@moduon MT-1075