[FIX] sale_packaging_default: setting product_uom_qty to 0 empties packaging
Before this patch, we had this problem:

1. Add new SO line.
2. Set a packaging and a packaging qty. The UoM qty is computed. Good.
3. Set UoM qty to zero. Packaging and packaging qty remain with previous value. Bad.

When user sets UoM to zero, Odoo should remove any packaging information.

Notice that, when adding a new line, that qty will be zero at the beginning, but we will want to still get the defaults. Thus, this change is only applied when the user *is* actually setting that value, and not on any other circumstance. I needed to override the `onchange()` method to propagate that information.

@moduon MT-3930