[IMP] sale_order_product_picker: quick add 1 packaging
When enabling packaging standard Odoo system, now the picker allows you to add +1 package.

For this to work, product packagings must be enabled in settings, and the product must have at least one packaging configured for sales.

Only the 1st one for sales will be used for the picker shortcut.

@moduon MT-4396

Co-authored-by: Rafael Blasco <8736623+rafaelbn@users.noreply.github.com>