[FIX] sale_order_product_recommendation: AccessError when getting recommendations
A simple salesperson couldn't access recommendations after https://github.com/OCA/sale-workflow/pull/2913 was merged.

It turns out that these tests were not using the testing user they created, although it seemed like so.

Now I'm reducing the permissions for the user that runs the tests, and thus can fix the issue and get it tested.

@moduon MT-4941