Apply rounding on discount
Fix computed fields dependencies

The issue with actual implementation is that discount is already
a computed field in std Odoo (eg to apply discount from pricelist).

Therefore, the value computed by Odoo must be stored in another field,
(ie discount1) for discount to be used to consolidate the multiple
discount. For this, we assign _compute_discount as the compute
of discount1 and redefine a new function to compute discount as
consolidated from the other discount fields.

Manually bump major version to play upgrade script

Fix migration script import

Fix init discount from pricelist