[FIX] product_supplierinfo_for_customer_sale: Don't inject context in onchange
The previous code was trying to avoid a double search on customerinfo
passing by context the already searched customerinfo, but the ORM makes
some artifacts (like duplicating one2many references with NewId records)
on that cases, provoking the call chain to fail when comparing .id on
NewId.

The solution is to not pass anything by context and do the search 2
times. Anyway, the performance improvement was low, having data in
cache.

TT49073