Added setting on partner
Added the ability to set a standard invoice policy for a partner so that it will always be sett on the sales order. It will select the default setting from Sales->Settings->Invoice->Invoice Policy as the default setting when a new partner is created.
Also added an extra menu item in the Action menu that can update all partners with the default setting if no setting is set from the beginning.