[IMP] sale_planner_calendar: Open links Whatsapp and Location in new window
Currently the Whatsapp and Location links open in the same window.
Once you click on one of these links, it overwrites the current window
and opens the whatsapp or google maps website.
When you go back, you have lost all context, you are back to the generic
routing screen. To avoid this, we have added the option to open the
links in a new window that can be closed and not lose the information
from the routing screen.

TT50045