[FIX] sale_order_line_input: New id for the _get_lang method
When a line is defined in the sales order it calculates the name of the
description with the language of the order using the new id of the
new order being created. If this is done from the order lines view there
is no order created and that is why you need to create a new id to
avoid the error in the parent method.