[IMP] sale_triple_discount: Improves perfs and consistency
When computing the amount total on a sale.order.line, we now avoid to play with the cache since this change introduced in https://github.com/OCA/sale-workflow/commit/e3e59ba15ebcf1f91124eda60a694c0a76fabb69 introduces a decline in performance. When the cache in invalidated, fields are flushed to the database and depending computed fields are recomputed..... This ultimately leads to temporary inconsistencies breaking others addons relaying on this one. Moreover, the method used in this original commit 'self.invalidate_cache(fnames=self._discount_fields(), ids=self.ids)' is deprecated and the use of the recommended new one lead to others side effects. The original approach is replaced by a contextual method modifying only the discount field when we need it's needed to don't break code relaying on the fact that the discount field is the total discount to apply to a line. This contextual method ensures that the original value is properly restored on exit and that it will not trigger additional recompute due to the temporary change of the discount value