[FIX] sale_sourced_by_line: fix conflict with precompute
Built-in filed is marked as precompute=True.
When field is marked as precomputed and readonly, Odoo will
remove it's value from create values. (See `odoo.models.BaseModel._prepare_create_values`)
Therefore, it was not possible to set `warehouse_id` in create
as it would reset to default computed warehouse.

Now, it's fixed and readonly enforced on view level.