[FIX] sale_order_secondary_unit: fix test
The behaviour of odoo in v17 when modifying an existing order line is to
recalculate the price based on the price set on the product and not on
the line. This module itself had an error in its tests as it was checking
a price based on the modified price of the line (1000) which was not
correct as when modifying any data of the order line the price was
recalculated based on the price of the product which by default is 1
if it has not been set.

TT54893