[IMP] sale_order_safe_commitment_date: delivery date helper
When the lead time has decimals (for example, 1.5 = 36 hours) a lot of false
positives are raised for the delivery exceptions. as the default hour for the
delivery date is the current one. For example: in that case if the current hour
is 01:00, the expected date will be 13:00 for the next day, but when we set the
delivery date for tomorrow, the default hour will be tomorrow at 1:00, which is
before that expected day. The users tend to ignore the warnings after stumbling
upon these cases over and over.

So now:

- There's a new delivery date helper field to simply set the date (no time) that
  sets the hour to 23.59.59 for the delivery date (commitment_date)
- The delivery date can be still changed manually whenever that info is important
  as usal in the standard.

MT-9790