[IMP] sale_order_line_sequence: Reduce column size through label + optional
Odoo, when auto-sizing the columns of a list view, takes the column
label for distributing the widths. Having a label "Line number" (or
longer in other languages), makes the column to take by default a lot of
space, while its contents will always be 1, 2, 3..., aka very sort
contents.

Let's change specifically in the view the label to "#" to have the
minimum size for the initial width distribution.

We take the occasion to make it optional (but shown by default) as well.

TT57583